# openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "API Documentation",
    "version" : "1.0"
  },
  "paths" : {
...
    "/api/auth/login" : {
      "post" : {
        "summary" : "login",
        "description" : "login",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "key" : {
                    "type" : "string"
                  }
                },
                "description" : ""
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "integer",
                      "description" : ""
                    },
                    "message" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "data" : {
                      "type" : "string",
                      "description" : ""
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/info" : {
      "get" : {
        "summary" : "getInfo",
        "description" : "getInfo",
        "parameters" : [ {
          "name" : "Authorization",
          "in" : "header",
          "description" : "",
          "required" : true,
          "schema" : { }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "integer",
                      "description" : ""
                    },
                    "message" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "data" : {
                      "type" : "object",
                      "properties" : {
                        "username" : {
                          "type" : "string"
                        },
                        "name" : {
                          "type" : "string"
                        },
                        "email" : {
                          "type" : "string"
                        },
                        "phone" : {
                          "type" : "string"
                        },
                        "id" : {
                          "type" : "integer"
                        }
                      },
                      "description" : ""
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/login" : {
      "post" : {
        "summary" : "login",
        "description" : "login",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "key" : {
                    "type" : "string"
                  }
                },
                "description" : ""
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "integer",
                      "description" : ""
                    },
                    "message" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "data" : {
                      "type" : "string",
                      "description" : ""
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/register" : {
      "post" : {
        "summary" : "register",
        "description" : "register",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "password" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "id" : {
                    "type" : "integer",
                    "description" : "",
                    "format" : "int64"
                  },
                  "username" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "realName" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "phone" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "email" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "status" : {
                    "type" : "integer",
                    "description" : "状态：0-禁用，1-启用"
                  },
                  "createdAt" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "updatedAt" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "avatarUrl" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "rolesStr" : {
                    "type" : "string",
                    "description" : "数据库中存储的字符串"
                  },
                  "roles" : {
                    "type" : "array",
                    "description" : "程序中使用的列表",
                    "items" : {
                      "type" : "string"
                    }
                  }
                },
                "description" : ""
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : { }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/info" : {
      "get" : {
        "summary" : "getInfo",
        "description" : "getInfo",
        "parameters" : [ {
          "name" : "Authorization",
          "in" : "header",
          "description" : "",
          "required" : true,
          "schema" : { }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "integer",
                      "description" : ""
                    },
                    "message" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "data" : {
                      "type" : "string",
                      "description" : ""
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put" : {
        "summary" : "updateInfo",
        "description" : "updateInfo",
        "parameters" : [ {
          "name" : "Authorization",
          "in" : "header",
          "description" : "",
          "required" : true,
          "schema" : { }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "password" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "id" : {
                    "type" : "integer",
                    "description" : "",
                    "format" : "int64"
                  },
                  "username" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "realName" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "phone" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "email" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "status" : {
                    "type" : "integer",
                    "description" : "状态：0-禁用，1-启用"
                  },
                  "createdAt" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "updatedAt" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "avatarUrl" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "rolesStr" : {
                    "type" : "string",
                    "description" : "数据库中存储的字符串"
                  },
                  "roles" : {
                    "type" : "array",
                    "description" : "程序中使用的列表",
                    "items" : {
                      "type" : "string"
                    }
                  }
                },
                "description" : ""
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "integer",
                      "description" : ""
                    },
                    "message" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "data" : {
                      "type" : "string",
                      "description" : ""
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/public/admin/register" : {
      "post" : {
        "summary" : "register",
        "description" : "register",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "username" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "password" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "email" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "phone" : {
                    "type" : "string",
                    "description" : ""
                  }
                },
                "description" : ""
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : { }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}